requirements.txt
pandas
datasets
tqdm
huggingface_hub
argparse
openai
python-dotenv
evaluate
bert_score
rouge_score